Refactor: reorganize multiSite functions and extract logic into separate files
